requirements.txt
web3>=6.0.0
requests
python-dotenv
colorama
python-telegram-bot
pyyaml
solana>=0.30.0
solders>=0.18.0
base58

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
bcrypt>=4.0.0
watchdog>=3.0.0
